% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bayesian_power.R
\name{analyze_bayesian}
\alias{analyze_bayesian}
\title{Analyze Virome Data Using Bayesian Methods}
\usage{
analyze_bayesian(counts, groups, prior_strength, credible_interval)
}
\arguments{
\item{counts}{Matrix of virome counts}

\item{groups}{Vector of group assignments}

\item{prior_strength}{Prior strength parameter}

\item{credible_interval}{Width of credible interval}
}
\value{
List with Bayesian analysis results
}
\description{
Internal function to perform Bayesian analysis on virome count data
}
\keyword{internal}
