% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bayesian_power.R
\name{analyze_bayesian_diversity}
\alias{analyze_bayesian_diversity}
\title{Analyze Alpha Diversity Using Bayesian Methods}
\usage{
analyze_bayesian_diversity(
  counts,
  groups,
  metrics,
  prior_strength = 2,
  credible_interval = 0.95,
  n_posterior_samples = 1000
)
}
\arguments{
\item{counts}{Matrix of virome counts}

\item{groups}{Vector of group assignments}

\item{metrics}{Character vector of alpha diversity metrics to analyze}

\item{prior_strength}{Prior strength parameter}

\item{credible_interval}{Width of credible interval}

\item{n_posterior_samples}{Number of posterior samples to generate}
}
\value{
List with Bayesian diversity analysis results
}
\description{
Internal function to perform Bayesian analysis on alpha diversity metrics
}
\keyword{internal}
