% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sample_size.R
\name{estimate_sample_size}
\alias{estimate_sample_size}
\title{Estimate Required Sample Size for Virome Studies}
\usage{
estimate_sample_size(
  power = 0.8,
  effect_size,
  n_viruses,
  alpha = 0.05,
  sparsity = 0.8,
  dispersion = 2,
  method = "wilcoxon"
)
}
\arguments{
\item{power}{Desired statistical power (default: 0.8)}

\item{effect_size}{Expected effect size (fold change)}

\item{n_viruses}{Number of viral taxa tested}

\item{alpha}{Significance level (default: 0.05)}

\item{sparsity}{Proportion of zeros in the data (default: 0.8)}

\item{dispersion}{Dispersion parameter for viral abundance (default: 2)}

\item{method}{Test method to use: "wilcoxon", "t.test", or "deseq" (default: "wilcoxon")}
}
\value{
A list containing estimated sample size and parameters
}
\description{
Estimates the number of samples needed to achieve a desired statistical power
for detecting differences in viral abundances between groups.
}
\examples{
sample_size <- estimate_sample_size(power = 0.8, effect_size = 1.5, n_viruses = 100)
}
